refactor: move the logic of signature_help to a new file
